+++ Packages/com.isaacbg77.gamejamkit/Runtime/Sound/IMusicService.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace GameJamKit.Sound
+{
+    public interface IMusicService
+    {
+        Awaitable PlayMusic(AudioClip clip);
+        Awaitable StopMusic();
+    }
+}