+++ Packages/com.isaacbg77.gamejamkit/Runtime/UI/IInitializableUI.cs
@@ -1,7 +1,0 @@
-namespace GameJamKit.UI
-{
-    public interface IInitializableUI
-    {
-        void Initialize();
-    }
-}