+++ Packages/com.isaacbg77.gamejamkit/Runtime/UI/ToMainMenu.cs
@@ -1,19 +1,18 @@
-using Eflatun.SceneReference;
 using GameJamKit.SceneManagement;
 using UnityEngine;
 
 namespace GameJamKit.UI.Menus
 {
-    public class CreditsMenu : MonoBehaviour
+    public class ToMainMenu : MonoBehaviour
     {
-        [SerializeField] private SceneReference? mainMenuScene;
+        [SerializeField] private SceneLoadData? mainMenuScene;
 
         private void Awake()
         {
-            if (mainMenuScene == null) Debug.LogError($"{nameof(CreditsMenu)}: {nameof(mainMenuScene)} is null");
+            if (mainMenuScene == null) Debug.LogError($"{nameof(ToMainMenu)}: {nameof(mainMenuScene)} is null");
         }
 
-        public void Back()
+        public void GoToMainMenu()
         {
             if (mainMenuScene == null) return;
             _ = SceneLoader.Instance.LoadSceneAsync(mainMenuScene);